refactor(workflow): separate WDLReader and NextflowReader into differnt files
